+++ packages/babel-compiler/package.js
@@ -6,11 +6,11 @@ Package.describe({
   // isn't possible because you can't publish a non-recommended
   // release with package versions that don't have a pre-release
   // identifier at the end (eg, -dev)
-  version: '6.4.0-modules.7'
+  version: '6.4.0-modules.8'
 });
 
 Npm.depends({
-  'meteor-babel': '0.7.1'
+  'meteor-babel': '0.7.2'
 });
 
 Package.onUse(function (api) {