+++ packages/check/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "Check whether a value matches a pattern",
-  version: '1.1.1-modules.7'
+  version: '1.1.1-modules.8'
 });
 
 Package.onUse(function (api) {
@@ -12,7 +12,7 @@ Package.onUse(function (api) {
 });
 
 Package.onTest(function (api) {
-  api.use(['check', 'tinytest', 'underscore', 'ejson'], ['client', 'server']);
+  api.use(['check', 'tinytest', 'underscore', 'ejson', 'ecmascript'], ['client', 'server']);
 
   api.addFiles('match_test.js', ['client', 'server']);
 });