+++ packages/meteor-env-dev/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'meteor-env-dev',
-  version: '0.0.1-modules.7',
+  version: '0.0.1-modules.8',
   summary: 'Package for setting up development-specific Meteor environment',
   debugOnly: true,
   documentation: 'README.md'