+++ packages/modules/package.js
@@ -1,8 +1,12 @@
 Package.describe({
   name: "modules",
-  version: "0.5.0-modules.7",
+  version: "0.5.0-modules.8",
   summary: "CommonJS module system",
   documentation: "README.md"
+});
+
+Npm.depends({
+  buffer: "4.4.0"
 });
 
 Package.onUse(function(api) {