+++ packages/test-in-console/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "Run tests noninteractively, with results going to the console.",
-  version: '1.0.8-modules.7'
+  version: '1.0.8-modules.8'
 });
 
 Package.onUse(function (api) {