+++ packages/url/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "Utility code for constructing URLs",
-  version: "1.0.6-modules.7"
+  version: "1.0.6-modules.8"
 });
 
 Package.onUse(function(api) {