+++ tools/isobuild/import-scanner.js
@@ -1,5 +1,5 @@
 import assert from "assert";
-import {isString, has, keys, each, without} from "underscore";
+import {isString, has, keys, each, map, without} from "underscore";
 import {sha1, readAndWatchFileWithHash} from "../fs/watch.js";
 import {matches as archMatches} from "../utils/archinfo.js";
 import {findImportedModuleIdentifiers} from "./js-analyze.js";
@@ -16,6 +16,17 @@ import {
   statOrNull,
   convertToPosixPath,
 } from "../fs/files.js";
+
+const nativeModulesMap = Object.create(null);
+
+Object.keys(process.binding("natives")).forEach(id => {
+  if (id === "freelist" ||
+      id.startsWith("internal/")) {
+    return;
+  }
+
+  nativeModulesMap[id] = true;
+});
 
 // Default handlers for well-known file extensions.
 const extensions = {
@@ -63,7 +74,8 @@ export default class ImportScanner {
     this.usedPackageNames = usedPackageNames;
     this.nodeModulesPath = nodeModulesPath;
     this.watchSet = watchSet;
-    this.absPathToOutputIndex = {};
+    this.absPathToOutputIndex = Object.create(null);
+    this.allMissingNodeModules = Object.create(null);
     this.outputFiles = [];
 
     this._statCache = new Map;
@@ -90,6 +102,10 @@ export default class ImportScanner {
       }
     });
 
+    return this;
+  }
+
+  scanImports() {
     this.outputFiles.forEach(file => {
       if (! file.lazy || file.imported) {
         this._scanFile(file);
@@ -100,6 +116,8 @@ export default class ImportScanner {
   }
 
   addNodeModules(identifiers) {
+    const newMissingNodeModules = Object.create(null);
+
     if (identifiers) {
       if (typeof identifiers === "object" &&
           ! Array.isArray(identifiers)) {
@@ -107,21 +125,45 @@ export default class ImportScanner {
       }
 
       if (identifiers.length > 0) {
-        this._scanFile({
-          sourcePath: "fake.js",
-          // By specifying the .deps property of this fake file ahead of
-          // time, we can avoid calling findImportedModuleIdentifiers in the
-          // _scanFile method.
-          deps: identifiers,
-        });
+        const previousAllMissingNodeModules = this.allMissingNodeModules;
+        this.allMissingNodeModules = newMissingNodeModules;
+
+        try {
+          this._scanFile({
+            sourcePath: "fake.js",
+            // By specifying the .deps property of this fake file ahead of
+            // time, we can avoid calling findImportedModuleIdentifiers in the
+            // _scanFile method.
+            deps: identifiers,
+          });
+
+        } finally {
+          this.allMissingNodeModules = previousAllMissingNodeModules;
+
+          // Remove previously seen missing module identifiers from
+          // newMissingNodeModules and merge the new identifiers back into
+          // this.allMissingNodeModules.
+          each(keys(newMissingNodeModules), key => {
+            if (has(previousAllMissingNodeModules, key)) {
+              delete newMissingNodeModules[key];
+            } else {
+              previousAllMissingNodeModules[key] =
+                newMissingNodeModules[key];
+            }
+          });
+        }
       }
     }
 
-    return this;
+    return newMissingNodeModules;
   }
 
   getOutputFiles(options) {
-    return this.outputFiles;
+    // Return all installable output files that are either eager or
+    // imported by another module.
+    return this.outputFiles.filter(file => {
+      return file.installPath && (! file.lazy || file.imported);
+    });
   }
 
   _findImportedModuleIdentifiers(file) {
@@ -160,10 +202,12 @@ export default class ImportScanner {
     }
 
     each(file.deps, id => {
-      const absImportedPath = this._tryToResolveImportedPath(file, id);
-      if (! absImportedPath) {
+      const resolved = this._tryToResolveImportedPath(file, id);
+      if (! resolved) {
         return;
       }
+
+      const absImportedPath = resolved.path;
 
       if (has(this.absPathToOutputIndex, absImportedPath)) {
         // Avoid scanning files that we've scanned before, but mark them
@@ -301,9 +345,8 @@ export default class ImportScanner {
     }
 
     const dirs = this._splitPath(pathDirname(installPath));
-    const bundlingClientApp =
-      ! this.name && // Indicates we are bundling an app.
-      archMatches(this.bundleArch, "web");
+    const isApp = ! this.name;
+    const bundlingForWeb = archMatches(this.bundleArch, "web");
 
     for (let dir of dirs) {
       if (dir.charAt(0) === "." ||
@@ -314,12 +357,21 @@ export default class ImportScanner {
         return;
       }
 
-      if (bundlingClientApp && (dir === "server" ||
-                                dir === "private")) {
-        // If we're bundling an app for a client architecture, any files
-        // contained by a server-only directory that is not contained by
-        // a node_modules directory must be ignored.
-        return;
+      if (isApp) {
+        if (bundlingForWeb) {
+          if (dir === "server" ||
+              dir === "private") {
+            // If we're bundling an app for a client architecture, any files
+            // contained by a server-only directory that is not contained by
+            // a node_modules directory must be ignored.
+            return;
+          }
+        } else if (dir === "client") {
+          // If we're bundling an app for a server architecture, any files
+          // contained by a client-only directory that is not contained by
+          // a node_modules directory must be ignored.
+          return;
+        }
       }
 
       if (dir === "node_modules") {
@@ -385,7 +437,7 @@ export default class ImportScanner {
       resolved = this._joinAndStat(dirPath, "index.js");
     }
 
-    return resolved && resolved.path;
+    return resolved;
   }
 
   _joinAndStat(...joinArgs) {
@@ -439,13 +491,23 @@ export default class ImportScanner {
   }
 
   _resolveNodeModule(file, id) {
-    let resolved = null;
-    let dir = pathJoin(this.sourceRoot, file.sourcePath);
+    const isNative = has(nativeModulesMap, id);
+    if (isNative && archMatches(this.bundleArch, "os")) {
+      // Forbid installing any server module with the same name as a
+      // native Node module.
+      return null;
+    }
 
-    do {
+    let dir = pathJoin(this.sourceRoot, file.sourcePath);
+    let resolved = this._joinAndStat(dir);
+    if (! resolved || ! resolved.stat.isDirectory()) {
       dir = pathDirname(dir);
-      resolved = this._joinAndStat(dir, "node_modules", id);
-    } while (! resolved && dir !== this.sourceRoot);
+    }
+
+    while (! (resolved = this._joinAndStat(dir, "node_modules", id)) &&
+           dir !== this.sourceRoot) {
+      dir = pathDirname(dir);
+    }
 
     if (! resolved && this.nodeModulesPath) {
       // After checking any local node_modules directories, fall back to
@@ -454,17 +516,13 @@ export default class ImportScanner {
     }
 
     if (! resolved) {
-      const parts = id.split("/");
-      if (parts[0] !== "meteor") { // Exclude meteor/... packages.
-        // If the imported identifier is neither absolute nor relative,
-        // but top-level, then it might be satisfied by a package
-        // installed in the top-level node_modules directory, and we
-        // should record the missing dependency so that we can include it
-        // in the app bundle.
-        const missing = file.missingNodeModules || Object.create(null);
-        missing[id] = true;
-        file.missingNodeModules = missing;
-      }
+      // If the imported identifier is neither absolute nor relative, but
+      // top-level, then it might be satisfied by a package installed in
+      // the top-level node_modules directory, and we should record the
+      // missing dependency so that we can include it in the app bundle.
+      const missing = file.missingNodeModules || Object.create(null);
+      this.allMissingNodeModules[id] = missing[id] = true;
+      file.missingNodeModules = missing;
     }
 
     // If the dependency is still not resolved, it might be handled by the
@@ -497,23 +555,46 @@ export default class ImportScanner {
   _resolvePkgJsonMain(dirPath, seenDirPaths) {
     const pkgJsonPath = pathJoin(dirPath, "package.json");
     const pkg = this._readPkgJson(pkgJsonPath);
+    if (! pkg) {
+      return null;
+    }
 
-    if (pkg && isString(pkg.main)) {
+    let main = pkg.main;
+
+    if (archMatches(this.bundleArch, "web") &&
+        isString(pkg.browser)) {
+      main = pkg.browser;
+    }
+
+    if (isString(main)) {
       // The "main" field of package.json does not have to begin with ./
       // to be considered relative, so first we try simply appending it to
       // the directory path before falling back to a full resolve, which
       // might return a package from a node_modules directory.
-      const resolved = this._joinAndStat(dirPath, pkg.main) ||
+      const resolved = this._joinAndStat(dirPath, main) ||
         // The _tryToResolveImportedPath method takes a file object as its
         // first parameter, but only the .sourcePath property is ever
         // used, so we can get away with passing a fake directory file
         // object with only that property.
         this._tryToResolveImportedPath({
-          sourcePath: dirPath,
-        }, pkg.main, seenDirPaths);
+          sourcePath: pathRelative(this.sourceRoot, dirPath),
+        }, main, seenDirPaths);
 
       if (resolved) {
-        this._addPkgJsonToOutput(pkgJsonPath, pkg);
+        // Output a JS module that exports just the "name", "version", and
+        // "main" properties defined in the package.json file.
+        const pkgSubset = {
+          name: pkg.name,
+        };
+
+        if (has(pkg, "version")) {
+          pkgSubset.version = pkg.version;
+        }
+
+        pkgSubset.main = main;
+
+        this._addPkgJsonToOutput(pkgJsonPath, pkgSubset);
+
         return resolved;
       }
     }
@@ -523,13 +604,9 @@ export default class ImportScanner {
 
   _addPkgJsonToOutput(pkgJsonPath, pkg) {
     if (! has(this.absPathToOutputIndex, pkgJsonPath)) {
-      const data = new Buffer(
-        // Output a JS module that exports just the "name", "version", and
-        // "main" properties defined in the package.json file.
-        "exports.name = " + JSON.stringify(pkg.name) + ";\n" +
-        "exports.version = " + JSON.stringify(pkg.version) + ";\n" +
-        "exports.main = " + JSON.stringify(pkg.main) + ";\n"
-      );
+      const data = new Buffer(map(pkg, (value, key) => {
+        return `exports.${key} = ${JSON.stringify(value)};\n`;
+      }).join(""));
 
       const relPkgJsonPath = pathRelative(this.sourceRoot, pkgJsonPath);