+++ tools/isobuild/package-source.js
@@ -1327,6 +1327,16 @@ _.extend(PackageSource.prototype, {
     const fileOptions = {};
     const dirs = files.pathDirname(relPath).split(files.pathSep);
 
+    // If the file is restricted to the opposite architecture, make sure
+    // it is not evaluated eagerly.
+    if (arch === "os") {
+      if (dirs.indexOf("client") >= 0) {
+        fileOptions.lazy = true;
+      }
+    } else if (dirs.indexOf("server") >= 0) {
+      fileOptions.lazy = true;
+    }
+
     if (dirs.indexOf("node_modules") >= 0) {
       fileOptions.lazy = true;
       fileOptions.transpile = false;