+++ tools/tests/apps/modules/packages/modules-test-package/package.js
@@ -5,8 +5,13 @@ Package.describe({
   documentation: 'README.md'
 });
 
+Npm.depends({
+  "os-browserify": "0.2.0"
+});
+
 Package.onUse(function(api) {
   api.use('ecmascript');
+  api.use('templating');
   api.mainModule("client.js", "client");
   api.mainModule("server.js", "server");
   api.export("ModulesTestPackage");